refactor: move vacuum command to operations module
