+++ rust/src/delta.rs
@@ -18,9 +18,8 @@ use super::table_state::DeltaTableState;
 use crate::action::{Add, Stats};
 use crate::delta_config::DeltaConfigError;
 use crate::storage::ObjectStoreRef;
-use crate::vacuum::{Vacuum, VacuumError};
 
-use chrono::{DateTime, Duration, Utc};
+use chrono::{DateTime, Utc};
 use futures::StreamExt;
 use lazy_static::lazy_static;
 use log::debug;
@@ -1082,24 +1081,6 @@ impl DeltaTable {
     /// metadata.
     pub fn get_min_writer_version(&self) -> i32 {
         self.state.min_writer_version()
-    }
-
-    /// Vacuum the delta table see [`Vacuum`] for more info
-    pub async fn vacuum(
-        &mut self,
-        retention_hours: Option<u64>,
-        dry_run: bool,
-        enforce_retention_duration: bool,
-    ) -> Result<Vec<String>, VacuumError> {
-        let mut plan = Vacuum::default()
-            .dry_run(dry_run)
-            .enforce_retention_duration(enforce_retention_duration);
-        if let Some(hours) = retention_hours {
-            plan = plan.with_retention_period(Duration::hours(hours as i64));
-        }
-
-        let res = plan.execute(self).await?;
-        Ok(res.files_deleted)
     }
 
     /// Return table schema parsed from transaction log. Return None if table hasn't been loaded or
